Task/creacion de metodo put
